+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- feat: only parse model when use is imported
+
 ## 1.0.0-beta.7
 
 ### Patch Changes